requirements.txt
# Core Dependencies
streamlit==1.32.0
plotly==5.18.0
pandas==2.2.0
numpy==1.24.3
python-dotenv==1.0.0

# AI & ML
openai==1.12.0
sentence-transformers==2.2.2
transformers==4.36.2
torch==2.2.0
scikit-learn==1.3.2

# Web & API
fastapi==0.109.0
uvicorn[standard]==0.27.0
httpx==0.26.0
requests==2.31.0
aiohttp==3.9.3
python-multipart==0.0.6

# Data Processing
beautifulsoup4==4.12.2
pydantic==2.5.3
trafilatura==1.6.1
markdown==3.5.2

# UI Components
streamlit-option-menu==0.3.6
pillow==10.2.0
altair==5.2.0

# Browser Automation
playwright==1.41.0

# Utilities
tqdm==4.66.1
pyyaml==6.0.1
jsonschema==4.21.1

# Optional - Development Only
pytest==8.0.0
black==24.1.1
flake8==7.0.0
mypy==1.8.0